Fix: Use dns.promises in mail-config to prevent callback error
The `mail-config` job crashes because it uses `await dns.resolveMx` which requires the Promise API, but the default import provides the Callback API.

Error observed: `The "callback" argument must be of type function. Received undefined`

This PR switches the import to `{ promises as dns }` to support the async/await usage present in the code.